requirements.txt
pandas==2.1.4
great_expectations 
requests
selenium
webdriver-manager
sqlalchemy
psycopg2-binary
python-dotenv
kaggle
apache-airflow[celery]==3.0.2
-c https://raw.githubusercontent.com/apache/airflow/constraints-3.0.2/constraints-3.9.txt